Split off resend entry and rename
